refactor: 拆分通知运行时职责
