+++ scripts/run_checks.py
@@ -16,7 +16,9 @@ COMPILE_TARGETS = (
     "goldmonitor/alert_rules.py",
     "goldmonitor/alert_log_runtime.py",
     "goldmonitor/alert_notification_runtime.py",
+    "goldmonitor/alert_delivery_runtime.py",
     "goldmonitor/daily_digest.py",
+    "goldmonitor/daily_digest_delivery_runtime.py",
     "goldmonitor/daily_digest_runtime.py",
     "goldmonitor/config_runtime.py",
     "goldmonitor/settings_runtime.py",
@@ -37,7 +39,9 @@ COMPILE_TARGETS = (
     "goldmonitor/market_data.py",
     "goldmonitor/market_runtime.py",
     "goldmonitor/news.py",
+    "goldmonitor/desktop_notification_runtime.py",
     "goldmonitor/notification_adapters.py",
+    "goldmonitor/notification_channel_runtime.py",
     "goldmonitor/notification_runtime.py",
     "goldmonitor/operations_runtime.py",
     "goldmonitor/data_archive.py",
