+++ tests/contract_checks.ps1
@@ -85,10 +85,10 @@ Assert-Contains -Path "static\app-state.js" -Pattern 'auth:\s*\{\s*token:\s*SOCK
 Assert-Contains -Path "goldmonitor\application.py" -Pattern 'APP_USER_MODEL_ID\s*=\s*"GoldMonitor\.App"' -Message "app.py must define the Windows AppUserModelID"
 Assert-Contains -Path "goldmonitor\application.py" -Pattern 'MACOS_LAUNCH_AGENT_ID' -Message "app.py must define the macOS LaunchAgent identifier"
 Assert-Contains -Path "goldmonitor\application.py" -Pattern 'create_macos_status_item' -Message "app.py must expose a macOS menu bar status item"
-Assert-Contains -Path "goldmonitor\notification_runtime.py" -Pattern 'display notification' -Message "notification runtime must support macOS notification center alerts"
-Assert-Contains -Path "goldmonitor\notification_runtime.py" -Pattern 'afplay' -Message "notification runtime must support macOS alert sounds"
+Assert-Contains -Path "goldmonitor\desktop_notification_runtime.py" -Pattern 'display notification' -Message "desktop notification runtime must support macOS notification center alerts"
+Assert-Contains -Path "goldmonitor\desktop_notification_runtime.py" -Pattern 'afplay' -Message "desktop notification runtime must support macOS alert sounds"
 Assert-Contains -Path "goldmonitor\application.py" -Pattern 'app_id=APP_USER_MODEL_ID' -Message "desktop toast must use the application AppUserModelID"
-Assert-Contains -Path "goldmonitor\notification_runtime.py" -Pattern 'icon=icon' -Message "desktop toast must include the app icon"
+Assert-Contains -Path "goldmonitor\desktop_notification_runtime.py" -Pattern 'icon=icon' -Message "desktop toast must include the app icon"
 Assert-Contains -Path "goldmonitor\socket_runtime.py" -Pattern '"close_choice": on_close_choice' -Message "socket runtime must receive close choices from the frontend"
 Assert-Contains -Path "goldmonitor\desktop_runtime.py" -Pattern 'emit\("show_close_dialog"' -Message "desktop runtime must ask the frontend to show close confirmation"
 Assert-Contains -Path "goldmonitor\application.py" -Pattern 'hide_main_window' -Message "app.py must hide the pywebview window to tray"
