+++ tests/test_run_checks_script.py
@@ -13,7 +13,9 @@ EXPECTED_COMPILE_TARGETS = (
     "goldmonitor/alert_rules.py",
     "goldmonitor/alert_log_runtime.py",
     "goldmonitor/alert_notification_runtime.py",
+    "goldmonitor/alert_delivery_runtime.py",
     "goldmonitor/daily_digest.py",
+    "goldmonitor/daily_digest_delivery_runtime.py",
     "goldmonitor/daily_digest_runtime.py",
     "goldmonitor/config_runtime.py",
     "goldmonitor/settings_runtime.py",
@@ -34,7 +36,9 @@ EXPECTED_COMPILE_TARGETS = (
     "goldmonitor/market_data.py",
     "goldmonitor/market_runtime.py",
     "goldmonitor/news.py",
+    "goldmonitor/desktop_notification_runtime.py",
     "goldmonitor/notification_adapters.py",
+    "goldmonitor/notification_channel_runtime.py",
     "goldmonitor/notification_runtime.py",
     "goldmonitor/operations_runtime.py",
     "goldmonitor/data_archive.py",
